Revert "moved dev mode options to three dots instead of share modal"
This reverts commit f8bf0cf2f627489ac5788a18e4682cfb5e1082fa.
