+++ src/components/PostControls/ShareMenu/ShareMenuItems.tsx
@@ -18,11 +18,13 @@ import {useDialogControl} from '#/components/Dialog'
 import {SendViaChatDialog} from '#/components/dms/dialogs/ShareViaChatDialog'
 import {ArrowOutOfBoxModified_Stroke2_Corner2_Rounded as ArrowOutOfBoxIcon} from '#/components/icons/ArrowOutOfBox'
 import {ChainLink_Stroke2_Corner0_Rounded as ChainLinkIcon} from '#/components/icons/ChainLink'
+import {Clipboard_Stroke2_Corner2_Rounded as ClipboardIcon} from '#/components/icons/Clipboard'
 import {PaperPlane_Stroke2_Corner0_Rounded as PaperPlaneIcon} from '#/components/icons/PaperPlane'
 import * as Menu from '#/components/Menu'
 import {useAgeAssurance} from '#/ageAssurance'
 import {useAnalytics} from '#/analytics'
 import {IS_IOS} from '#/env'
+import {useDevMode} from '#/storage/hooks/dev-mode'
 import {RecentChats} from './RecentChats'
 import {type ShareMenuItemsProps} from './ShareMenuItems.types'
 
@@ -35,6 +37,7 @@ let ShareMenuItems = ({
   const {_} = useLingui()
   const navigation = useNavigation<NavigationProp>()
   const sendViaChatControl = useDialogControl()
+  const [devModeEnabled] = useDevMode()
   const aa = useAgeAssurance()
 
   const postUri = post.uri
@@ -76,6 +79,16 @@ let ShareMenuItems = ({
       conversation,
       embed: postUri,
     })
+  }
+
+  const onCopyATURI = async () => {
+    await ExpoClipboard.setStringAsync(postUri)
+    Toast.show(_(msg`Copied to clipboard`), 'clipboard-check')
+  }
+
+  const onCopyAuthorDID = async () => {
+    await ExpoClipboard.setStringAsync(postAuthor.did)
+    Toast.show(_(msg`Copied to clipboard`), 'clipboard-check')
   }
 
   return (
@@ -132,6 +145,29 @@ let ShareMenuItems = ({
                 <Trans>This post is only visible to logged-in users.</Trans>
               </Admonition>
             </Menu.ContainerItem>
+          </Menu.Group>
+        )}
+
+        {devModeEnabled && (
+          <Menu.Group>
+            <Menu.Item
+              testID="postAtUriCopyBtn"
+              label={_(msg`Copy post at:// URI`)}
+              onPress={onCopyATURI}>
+              <Menu.ItemText>
+                <Trans>Copy post at:// URI</Trans>
+              </Menu.ItemText>
+              <Menu.ItemIcon icon={ClipboardIcon} position="right" />
+            </Menu.Item>
+            <Menu.Item
+              testID="postAuthorDIDCopyBtn"
+              label={_(msg`Copy author DID`)}
+              onPress={onCopyAuthorDID}>
+              <Menu.ItemText>
+                <Trans>Copy author DID</Trans>
+              </Menu.ItemText>
+              <Menu.ItemIcon icon={ClipboardIcon} position="right" />
+            </Menu.Item>
           </Menu.Group>
         )}
       </Menu.Outer>
